# docs/swagger.yaml
definitions:
...
  controllers.ChangePasswordInput:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  controllers.RegisPublisherInput:
    properties:
      logo_url:
        type: string
    type: object
...
  controllers.ReviewInput:
    properties:
      content:
        type: string
      rate:
        type: integer
    type: object
...
  models.Game:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      genre_id:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      publisher_id:
        type: integer
      ratings:
        type: integer
      ratings_counter:
        type: integer
      release_date:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Review:
    properties:
      content:
        type: string
      created_at:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      rate:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /change-password:
    patch:
      description: Used by User, Publisher, or Admin for change password account.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to change user password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Change password account.
      tags:
      - Authentication & Authorization
...
  /games/{id}/review:
    post:
      description: Create new review and rate(1-5)
      parameters:
      - description: the body to create new review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ReviewInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Add review into this game
      tags:
      - Games
  /games/{id}/reviews:
    get:
      description: Get all reviews of spesific games
      parameters:
      - description: Game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Show all review of specific games by game_id
      tags:
      - Public
...
  /profiles:
    get:
      description: Get user, publisher, or admin info
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Get info of current login account
      tags:
      - Authentication & Authorization
...
  /publishers/{id}/games:
    get:
      description: Get all games of spesific publisher by id
      parameters:
      - description: Publisher Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
      summary: Show all game published by specific publisher
      tags:
      - Public
  /regist-admin:
    patch:
      description: Create an account before upgrade your role. (temporary) User or
        Admin can change into this role
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to become admin
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisPublisherInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Change role to become admin.
      tags:
      - Authentication & Authorization
  /regist-publisher:
    patch:
      description: Create an account before upgrade your role. (temporary) User or
        Admin can change into this role
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to become publisher
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisPublisherInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Change role to become publisher.
      tags:
      - Authentication & Authorization
...
  /users/games/reviews/{id}:
    delete:
      description: Only user who create this review have permission to update
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete review created by this user
      tags:
      - Users
    patch:
      description: Only user who create this review have permission to update
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review Id
        in: path
        name: id
        required: true
        type: string
      - description: the body to create new review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ReviewInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Update review created by this user
      tags:
      - Users
...
swagger: "2.0"